fix: handle larger arrays in take (#1383)
Closes #1382

TODO: 
* [x] Test with a dataset that has 5 billion values

+++ python/src/dataset.rs
@@ -415,7 +415,7 @@ impl Dataset {
 
     fn take(
         self_: PyRef<'_, Self>,
-        row_indices: Vec<usize>,
+        row_indices: Vec<u64>,
         columns: Option<Vec<String>>,
     ) -> PyResult<PyObject> {
         let projection = if let Some(columns) = columns {
